Move test step to the test class
